Changes to RVSWT101 Final
Correct the devices arguments for the BLE instrument.
Add commands to exercise the actuators and probe if the UUT is in
position.
Remove the No button pressed measurment.
Handle error when unpacking the bluetooth payload.
Add more actions to device.RVSWT101:reset().
Specify different button code limits for 4 and 6 button models.
Update final test to not prompt ui_buttonpress.
Check the bt scan_count matches the number of UUT buttons.
Add prompts to add or remove UUT to/from the fixture.